+++ CHANGELOG.md
@@ -3,4 +3,8 @@
 * Bugfix in volume-writing location, and force pretty formatting on written logs
 * Made the existing test actually query the node it created
 * Added another test to demonstrate an initial network setup
+* Adding copyright headers
+* Renamed tests to have more descriptive names
+* When asked about test suite data, send back a JSON of test suite metadata (rather than just a list of test names)
+* Made log level configurable
 * Add CircleCI